Fix: handle not JSON content in response parsing
Some intuit API as invoice download return not JSON content (PDF in this
case). `makeApiCall` wasn't working with it because of mandatory
response body parsing.

So if the response is not JSON, we don't want to parse the body.
And simply let the caller decide what to do with it.